% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{bottom_openings_industry}
\alias{bottom_openings_industry}
\title{bottom openings information
Input the dataframe about career and return the industry with bottom openings}
\usage{
bottom_openings_industry(data)
}
\description{
bottom openings information
Input the dataframe about career and return the industry with bottom openings
}
