% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{bottom_openings_job}
\alias{bottom_openings_job}
\title{bottom openings job
Input the dataframe about career and return the job with bottom openings}
\usage{
bottom_openings_job(data)
}
\description{
bottom openings job
Input the dataframe about career and return the job with bottom openings
}
