% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{bottom_wage_industry}
\alias{bottom_wage_industry}
\title{bottom wage information
Input the dataframe about career and return the job with bottom wage}
\usage{
bottom_wage_industry(data)
}
\description{
bottom wage information
Input the dataframe about career and return the job with bottom wage
}
