% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{bottom_wage_job}
\alias{bottom_wage_job}
\title{bottom wage job
Input the dataframe about career and return the job with bottom wage}
\usage{
bottom_wage_job(data)
}
\description{
bottom wage job
Input the dataframe about career and return the job with bottom wage
}
