% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{check_package}
\alias{check_package}
\title{check_package
use it to check you package is installed successfully}
\usage{
check_package()
}
\description{
check_package
use it to check you package is installed successfully
}
