% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_figure.R
\name{confirm_columns}
\alias{confirm_columns}
\title{Confirm Columns}
\usage{
confirm_columns(input_data, colname)
}
\arguments{
\item{...}{Column name - this must be in the form of a string.}
}
\value{
If all inputs are valid the output will produce a file which can be passed to \code{\link{plot_histogram}.
If inputs are invalid, an error message will be generated.
}
\description{
\code{confirm_columns} Checks correct columns exist within dataframe.
}
\details{

}
\examples{
confirm_columns("supp_table4.csv")

\dontrun{
confirm_columns(column, "Missing column")
}
}
