% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firingr8-package.R
\docType{package}
\name{firingr8-package}
\alias{firingr8}
\alias{firingr8-package}
\title{firingr8: Counts Firing Rates from Electrophysiology CSV Data}
\description{
Counts the firing rates for each sweep from Patch-Clamp Electrophysiology Data. Also gets various data electrophysiologists may be interested in such as half width, resting membrane potential, peak voltage, peak to peak frequency, interevent interval, and many others.
}
\keyword{internal}
