% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower_analyze.R
\name{flower_analyze}
\alias{flower_analyze}
\title{flower analyze}
\usage{
flower_analyze(input, group, data)
}
\arguments{
\item{input}{the response variable}

\item{group}{the predictor variable or the treatment grouping}

\item{data}{the data frame to be described}
}
\description{
preforms an ANOVA and returns summary of results
}
\examples{
flower_analyze(flowers$inflorescence, flowers$nitrogen, flowers)
}
\keyword{ANOVA,}
\keyword{summary}
