% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower_describe.R
\name{flower_describe}
\alias{flower_describe}
\title{flower describe}
\usage{
flower_describe(data, groupby, interest)
}
\arguments{
\item{data}{the data frame to be described}

\item{groupby}{the treatment grouping}

\item{interest}{the response variable}
}
\description{
groups data by experimental treatment and returns summary statistics
}
\examples{
flower_describe(flowers, flowers$nitrogen, flowers$inflorescence)
}
\keyword{describe,}
\keyword{summary}
