% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower_visualize.R
\name{flower_visualize}
\alias{flower_visualize}
\title{flower visualize}
\usage{
flower_visualize(input, group, data)
}
\arguments{
\item{input}{the response variable}

\item{group}{the predictor variable or the treatment grouping}

\item{data}{the data frame to be visualized}
}
\description{
visualizes data as a boxplot
}
\examples{
flower_visualize(flowers$inflorescence, flowers$nitrogen, flowers)
}
\keyword{boxplot}
\keyword{visualize,}
