% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{get_engineering_jobs}
\alias{get_engineering_jobs}
\title{get engineering jobs' information
process the jobs about engineering and plot the distrbution in a scatter figure}
\usage{
get_engineering_jobs(data)
}
\description{
get engineering jobs' information
process the jobs about engineering and plot the distrbution in a scatter figure
}
