% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{get_liberal_jobs}
\alias{get_liberal_jobs}
\title{get liberal jobs' information
process the liberal jobs and plot the distrbution in a scatter figure}
\usage{
get_liberal_jobs(data)
}
\description{
get liberal jobs' information
process the liberal jobs and plot the distrbution in a scatter figure
}
