% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{get_service_jobs}
\alias{get_service_jobs}
\title{get service jobs' information
process the service jobs and plot the distrbution in a scatter figure}
\usage{
get_service_jobs(data)
}
\description{
get service jobs' information
process the service jobs and plot the distrbution in a scatter figure
}
