% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{get_traditional_jobs}
\alias{get_traditional_jobs}
\title{get traditional jobs' information
process the liberal jobs and plot the distrbution in a scatter figure}
\usage{
get_traditional_jobs(data)
}
\description{
get traditional jobs' information
process the liberal jobs and plot the distrbution in a scatter figure
}
