% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{load_career_data}
\alias{load_career_data}
\title{load dataset
The function will return the dataframe about career}
\usage{
load_career_data()
}
\description{
load dataset
The function will return the dataframe about career
}
