% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{mean_wage}
\alias{mean_wage}
\title{mean wage
Input the dataframe about career and return the mean wage}
\usage{
mean_wage(data)
}
\description{
mean wage
Input the dataframe about career and return the mean wage
}
