% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{minimum_n_per_group: Minimum sample size per group for two-sample T-test.}
\usage{
minimum_n_per_group(d, power = 0.8, sig.level = 0.05)
}
\arguments{
\item{d}{The expected Cohen's d effect size.}

\item{power}{The desired statistical power (default is 0.8).}

\item{sig.level}{The desired significance level (default is 0.05).}
}
\value{
The minimum sample size per group needed for a two-sample t-test.
}
\description{
This function calculates the minimum sample size per group needed for a two-sample t-test,
given an expected Cohen's d effect size and desired statistical power.
}
\examples{
d = 0.5
minimum_n_per_group(d)
}
