% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_figure.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot histogram}
\usage{
plot_histogram(input_data, colname, binwidth = 0.5, xh = 10, yh = 160)
}
\arguments{
\item{...}{The column name which should be a string}
}
\value{
If all inputs are valid the output will produce a simple histogram.
If inputs are invalid, an error message will be generated.
}
\description{
\code{plot_histogram} Plots a pre-specified histogram powered by ggplot.
}
\details{

}
\examples{
plot_histogram(newst, "Avg.lifetime", binwidth=0.5, xh=10, yh=160)

}
