% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firingr8.R
\name{process_ephys_csv}
\alias{process_ephys_csv}
\title{A function for processing pClamp voltage trace data}
\usage{
process_ephys_csv(csvfile_path, experiment_identifier_path, output_path)
}
\arguments{
\item{csvfile_path}{Path to the data that needs to be processed}

\item{experiment_identifier_path}{Path to the file-identifier csv}

\item{output_path}{Path to the folder where the processed and visualized data will be saved}
}
\description{
This function allows you to detect action potentials and extract membrane properties in a batch file processing format
}
\examples{
process_ephys_csv(csvfile_path = "./data/raw_ephys_data.csv", experiment_identifier_path = "./data/identifiers_sample.csv", output_path = ".")
}
\keyword{electrophysiology}
