% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{show_data}
\alias{show_data}
\title{show dataset function
you can use it to look the deatailed data in dataframe}
\usage{
show_data(data)
}
\description{
show dataset function
you can use it to look the deatailed data in dataframe
}
