% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{sum_openings}
\alias{sum_openings}
\title{sum job openings
Input the dataframe about career and return the sum of openings}
\usage{
sum_openings(data)
}
\description{
sum job openings
Input the dataframe about career and return the sum of openings
}
