% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{top_openings_industry}
\alias{top_openings_industry}
\title{top openings information
Input the dataframe about career and return the industry with top openings}
\usage{
top_openings_industry(data)
}
\description{
top openings information
Input the dataframe about career and return the industry with top openings
}
