% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{top_openings_job}
\alias{top_openings_job}
\title{top openings job
Input the dataframe about career and return job with top openings}
\usage{
top_openings_job(data)
}
\description{
top openings job
Input the dataframe about career and return job with top openings
}
