% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{top_wage_industry}
\alias{top_wage_industry}
\title{top wage information
Input the dataframe about career and return the job with top wage}
\usage{
top_wage_industry(data)
}
\description{
top wage information
Input the dataframe about career and return the job with top wage
}
