% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.R
\name{top_wage_job}
\alias{top_wage_job}
\title{top wage job
Input the dataframe about career and return the job with top wage}
\usage{
top_wage_job(data)
}
\description{
top wage job
Input the dataframe about career and return the job with top wage
}
